moved functorfilter instances
