+++ modules/aql/arrow-query-language/src/main/kotlin/arrow/aql/extensions/list.kt
@@ -14,7 +14,7 @@ import arrow.extension
 import arrow.core.extensions.listk.applicative.applicative
 import arrow.core.extensions.listk.foldable.foldable
 import arrow.core.extensions.listk.functor.functor
-import arrow.mtl.extensions.listk.functorFilter.functorFilter
+import arrow.core.extensions.listk.functorFilter.functorFilter
 import arrow.typeclasses.FunctorFilter
 import arrow.typeclasses.Applicative
 import arrow.typeclasses.Foldable
