+++ modules/aql/arrow-query-language/src/main/kotlin/arrow/aql/extensions/option.kt
@@ -14,7 +14,7 @@ import arrow.extension
 import arrow.core.extensions.option.applicative.applicative
 import arrow.core.extensions.option.foldable.foldable
 import arrow.core.extensions.option.functor.functor
-import arrow.mtl.extensions.option.functorFilter.functorFilter
+import arrow.core.extensions.option.functorFilter.functorFilter
 import arrow.typeclasses.FunctorFilter
 import arrow.typeclasses.Applicative
 import arrow.typeclasses.Foldable
