+++ modules/aql/arrow-query-language/src/main/kotlin/arrow/aql/extensions/try.kt
@@ -14,7 +14,7 @@ import arrow.extension
 import arrow.core.extensions.`try`.applicative.applicative
 import arrow.core.extensions.`try`.foldable.foldable
 import arrow.core.extensions.`try`.functor.functor
-import arrow.mtl.extensions.`try`.functorFilter.functorFilter
+import arrow.core.extensions.`try`.functorFilter.functorFilter
 import arrow.typeclasses.FunctorFilter
 import arrow.typeclasses.Applicative
 import arrow.typeclasses.Foldable
