+++ modules/core/arrow-core/src/main/kotlin/arrow/core/extensions/TryFunctorFilter.kt
@@ -1,4 +1,4 @@
-package arrow.mtl.extensions
+package arrow.core.extensions
 
 import arrow.Kind
 import arrow.core.ForTry
